move time formatting to time util
